+++ tests/test_two_databases.py
@@ -30,7 +30,7 @@ class TestTwoDatabases(test.SimpleTestCase):
         with self.assertRaises(OperationalError):
             await self.db.execute_query("SELECT * FROM eventtwo")
 
-        results = await self.second_db.execute_query("SELECT * FROM eventtwo")
+        _, results = await self.second_db.execute_query("SELECT * FROM eventtwo")
         self.assertEqual(dict(results[0]), {"id": 1, "name": "Event", "tournament_id": 1})
 
     async def test_two_databases_relation(self):
@@ -40,7 +40,7 @@ class TestTwoDatabases(test.SimpleTestCase):
         with self.assertRaises(OperationalError):
             await self.db.execute_query("SELECT * FROM eventtwo")
 
-        results = await self.second_db.execute_query("SELECT * FROM eventtwo")
+        _, results = await self.second_db.execute_query("SELECT * FROM eventtwo")
         self.assertEqual(dict(results[0]), {"id": 1, "name": "Event", "tournament_id": 1})
 
         teams = []