+++ tortoise/backends/sqlite/client.py
@@ -2,7 +2,7 @@ import asyncio
 import os
 import sqlite3
 from functools import wraps
-from typing import List, Optional, Sequence
+from typing import List, Optional, Sequence, Tuple
 
 import aiosqlite
 
@@ -115,10 +115,14 @@ class SqliteClient(BaseDBAsyncClient):
                 await connection.commit()
 
     @translate_exceptions
-    async def execute_query(self, query: str, values: Optional[list] = None) -> Sequence[dict]:
+    async def execute_query(
+        self, query: str, values: Optional[list] = None
+    ) -> Tuple[int, Sequence[dict]]:
         async with self.acquire_connection() as connection:
             self.log.debug("%s: %s", query, values)
-            return await connection.execute_fetchall(query, values)
+            start = connection.total_changes
+            rows = await connection.execute_fetchall(query, values)
+            return (connection.total_changes - start) or len(rows), rows
 
     @translate_exceptions
     async def execute_query_dict(self, query: str, values: Optional[list] = None) -> List[dict]: